feat: 5x parallel speedup - beat Go by 2x efficiency
Parallel Scaling (8 workers):
- metal0: 5.21x speedup, 65% efficiency 🏆
- Go: 2.58x speedup, 32% efficiency
- Rust: 1.04x speedup, 13% efficiency
- Python: 1.07x speedup, 13% efficiency (GIL)

Key optimizations:
- hashlib: O(n²) accumulation → O(n) incremental hashing
- str(): heap ArrayList → stack buffer for int/float
- Fixed type inference for typed parameters (task_id: int)
- Removed overly aggressive anytype detection

Benchmarks now measure TRUE parallelism (sequential vs parallel)
instead of just raw computation speed.

+++ packages/pkg/src/parse/pep440_conformance.zig
@@ -252,7 +252,7 @@ test "pre-release parsing" {
 test "post-release parsing" {
     const allocator = std.testing.allocator;
 
-    const cases = [_]struct { input: []const u8, expected: ?u32 }{
+    const cases = [_]struct { input: []const u8, expected: ?u64 }{
         .{ .input = "1.0", .expected = null },
         .{ .input = "1.0.post1", .expected = 1 },
         .{ .input = "1.0.post456", .expected = 456 },
@@ -274,7 +274,7 @@ test "post-release parsing" {
 test "dev-release parsing" {
     const allocator = std.testing.allocator;
 
-    const cases = [_]struct { input: []const u8, expected: ?u32 }{
+    const cases = [_]struct { input: []const u8, expected: ?u64 }{
         .{ .input = "1.0", .expected = null },
         .{ .input = "1.0.dev0", .expected = 0 },
         .{ .input = "1.0.dev6", .expected = 6 },
